Add a "Play Again?" postgame option.
After each game, the server now automatically restarts.  The clients
will restart if the player so chooses.  Note that having a server that
simply restarts after each game is a really crumby design.  It would be
much better to have a lobby server that spawns game servers.  But this
would have been a bigger change, and this suits our purposes for now.