Try (unsuccessfully) to implement fog of war.
I'm pretty sure the problem is that stencil mode is enabled (when the
stencil group is drawn) but is never disabled (because there are
initially no armies to draw).  A more robust design would be to have a
parent "stencil" group and child groups for the clipping masks and
clipped objects.