Upgrade the game engine.
Armed cat is not totally functional with the new game engine, but the
remaining bugs are minor and do not correspond to major issues with the
engine.  Since we're planning to remove most of the armed cat features
anyways, I didn't put too much effort it making everything work again.